+++ example/lib/main.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:superdeck/models/options_model.dart';
 import 'package:superdeck/superdeck.dart';
 
 import 'src/style.dart';
 import 'src/widget/mix_demo.dart';
 
 void main() async {
+  await SuperDeckApp.initialize();
   runApp(
     Builder(builder: (context) {
       return MaterialApp(