+++ lib/helpers/layout_builder.dart
@@ -3,10 +3,9 @@ import 'package:flutter/material.dart';
 import '../components/atoms/image_widget.dart';
 import '../components/molecules/code_preview.dart';
 import '../components/molecules/slide_content.dart';
-import '../models/options_model.dart';
-import '../models/slide_model.dart';
 import '../providers/slide_provider.dart';
 import '../superdeck.dart';
+import 'constants.dart';
 import 'measure_size.dart';
 
 abstract class SlideBuilder<T extends Slide> extends StatelessWidget {
@@ -147,13 +146,34 @@ class ImageSlideBuilder extends SplitSlideBuilder<ImageSlide> {
     final src = config.options.src;
     final boxFit = config.options.fit?.toBoxFit() ?? spec.image.fit;
 
+    // THis slide breaks in half and I want to calculate the size based on if its in top or bottom
+    // or left or right. Also there is a property called flex which is how much of the slide it takes
+    // so I can use that to calculate the size of the canvas
+    final firstHalf =
+        config.contentOptions?.flex ?? const ContentOptions().flex;
+    final imageHalf = config.options.flex;
+
+// available size const width = 1280.0;
+//const height = 720.0;
+
+    double width;
+    double height;
+    const availableSize = kResolution;
+    if (config.options.position.isHorizontal()) {
+      width = availableSize.width * firstHalf / (firstHalf + imageHalf);
+      height = availableSize.height;
+    } else {
+      width = availableSize.width;
+      height = availableSize.height * firstHalf / (firstHalf + imageHalf);
+    }
+
     final side = Container(
       height: spec.image.height,
       width: spec.image.width,
       alignment: spec.image.alignment,
       decoration: BoxDecoration(
         image: DecorationImage(
-          image: getImageProvider(src),
+          image: getImageProvider(src, Size(width, height)),
           centerSlice: spec.image.centerSlice,
           repeat: spec.image.repeat ?? ImageRepeat.noRepeat,
           filterQuality: spec.image.filterQuality ?? FilterQuality.low,